Major code change - Changed PK from int to long for most of the models

+++ RideShareApp/app/src/main/java/com/digitusrevolution/rideshare/activity/HomePageActivity.java
@@ -149,8 +149,8 @@ public class HomePageActivity extends BaseActivity
         mProfilePhotoImageView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String GET_USER_PROFILE = APIUrl.GET_USER_PROFILE.replace(APIUrl.SIGNEDIN_USER_ID_KEY, Integer.toString(mUser.getId()))
-                        .replace(APIUrl.USER_ID_KEY, Integer.toString(mUser.getId()));
+                String GET_USER_PROFILE = APIUrl.GET_USER_PROFILE.replace(APIUrl.SIGNEDIN_USER_ID_KEY, Long.toString(mUser.getId()))
+                        .replace(APIUrl.USER_ID_KEY, Long.toString(mUser.getId()));
                 mCommonUtil.showProgressDialog();
                 RESTClient.get(GET_USER_PROFILE, null, new RSJsonHttpResponseHandler(mCommonUtil) {
                     @Override
